+++ sconnData/sconnConnector/POCO/Config/sconn/Gsm/sconnGsmRcpt.cs
@@ -11,7 +11,7 @@ namespace sconnConnector.POCO.Config.sconn
 {
     public class sconnGsmRcpt : ISerializableConfiguration,  IFakeAbleConfiguration
     {
-        public int Id { get; set; }
+        public ushort Id { get; set; }
 
         public string Name { get; set; }
