+++ sconnData/sconnConnector/POCO/Config/sconn/IO/sconnOutput.cs
@@ -41,8 +41,8 @@ namespace sconnConnector.POCO.Config
         public bool Value { get; set; }
         public bool Enabled { get; set; }
         public string Name { get; set; }
-        public byte NameId { get; set; }
-        public int ZoneId { get; set; }
+        public ushort NameId { get; set; }
+        public ushort ZoneId { get; set; }
         public DeviceIoCategory IoCategory { get; set; }
 
         private static Logger _logger = LogManager.GetCurrentClassLogger();
